Rename Layer helpers, consolidate noise tests, remove empty private section, and restore CMakeLists
