--- test/src/NoiseStyleTest.cpp
@@ -1,315 +1,0 @@
-/////////////////////////////////////////////////////////////////////////////////////////////////
-//
-//  Tencent is pleased to support the open source community by making tgfx available.
-//
-//  Copyright (C) 2026 Tencent. All rights reserved.
-//
-//  Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
-//  in compliance with the License. You may obtain a copy of the License at
-//
-//      https://opensource.org/licenses/BSD-3-Clause
-//
-//  unless required by applicable law or agreed to in writing, software distributed under the
-//  license is distributed on an "as is" basis, without warranties or conditions of any kind,
-//  either express or implied. see the license for the specific language governing permissions
-//  and limitations under the license.
-//
-/////////////////////////////////////////////////////////////////////////////////////////////////
-
-#include "tgfx/core/Image.h"
-#include "tgfx/core/PictureRecorder.h"
-#include "tgfx/layers/DisplayList.h"
-#include "tgfx/layers/ShapeLayer.h"
-#include "tgfx/layers/ShapeStyle.h"
-#include "tgfx/layers/SolidLayer.h"
-#include "tgfx/layers/TextLayer.h"
-#include "tgfx/layers/layerstyles/DropShadowStyle.h"
-#include "tgfx/layers/layerstyles/NoiseStyle.h"
-#include "tgfx/layers/vectors/ImagePattern.h"
-#include "utils/TestUtils.h"
-
-namespace tgfx {
-
-TGFX_TEST(NoiseStyleTest, NoiseShiftText) {
-  ContextScope scope;
-  auto context = scope.getContext();
-  ASSERT_TRUE(context != nullptr);
-  auto surface = Surface::Make(context, 1000, 1000);
-  ASSERT_TRUE(surface != nullptr);
-  auto displayList = std::make_unique<DisplayList>();
-  displayList->setRenderMode(RenderMode::Tiled);
-  displayList->setTileSize(64);
-
-  auto back = SolidLayer::Make();
-  back->setColor(Color::White());
-  back->setWidth(1000);
-  back->setHeight(1000);
-
-  auto typeface = MakeTypeface("resources/font/NotoSansSC-Regular.otf");
-  ASSERT_TRUE(typeface != nullptr);
-
-  auto noise = NoiseStyle::MakeMono(10.0f, 0.5f, Color::FromRGBA(0, 0, 0, 128), 42.0f);
-
-  auto textLayer = TextLayer::Make();
-  textLayer->setText("style");
-  textLayer->setFont(Font(typeface, 100.f));
-  textLayer->setTextColor(Color::FromRGBA(192, 192, 192, 255));
-  textLayer->setLayerStyles({noise});
-  back->addChild(textLayer);
-  displayList->root()->addChild(back);
-
-  for (int i = 0; i < 8; i++) {
-    surface->getCanvas()->clear();
-    textLayer->setMatrix(Matrix::MakeTrans(-50.f + static_cast<float>(i) * 20.f, 20.f));
-    displayList->render(surface.get());
-    EXPECT_TRUE(
-        Baseline::Compare(surface, "NoiseStyleTest/NoiseShiftText_frame" + std::to_string(i)));
-  }
-
-  auto textBounds = textLayer->getBounds();
-  float centerX = (textBounds.left + textBounds.right) * 0.5f;
-  float centerY = (textBounds.top + textBounds.bottom) * 0.5f;
-  float lastX = -50.f + 7 * 20.f;
-  float lastY = 20.f;
-
-  for (int i = 0; i < 5; i++) {
-    surface->getCanvas()->clear();
-    float angle = static_cast<float>(i + 1) * 20.f;
-    auto matrix = Matrix::MakeTrans(lastX + centerX, lastY + centerY) * Matrix::MakeRotate(angle) *
-                  Matrix::MakeTrans(-centerX, -centerY);
-    textLayer->setMatrix(matrix);
-    displayList->render(surface.get());
-    EXPECT_TRUE(
-        Baseline::Compare(surface, "NoiseStyleTest/NoiseShiftText_frame" + std::to_string(i + 8)));
-  }
-}
-
-TGFX_TEST(NoiseStyleTest, NoiseShapeText) {
-  ContextScope scope;
-  auto context = scope.getContext();
-  ASSERT_TRUE(context != nullptr);
-  auto surface = Surface::Make(context, 1000, 1000);
-  ASSERT_TRUE(surface != nullptr);
-  auto displayList = std::make_unique<DisplayList>();
-  displayList->setRenderMode(RenderMode::Tiled);
-  displayList->setTileSize(64);
-
-  auto back = SolidLayer::Make();
-  back->setColor(Color::White());
-  back->setWidth(1000);
-  back->setHeight(1000);
-
-  auto typeface = MakeTypeface("resources/font/NotoSansSC-Regular.otf");
-  ASSERT_TRUE(typeface != nullptr);
-
-  auto noise = NoiseStyle::MakeMono(10.0f, 0.5f, Color::FromRGBA(0, 0, 0, 128), 42.0f);
-
-  auto shapeLayer = ShapeLayer::Make();
-  Path path = {};
-  path.addRect(Rect::MakeXYWH(50, 50, 400, 150));
-  shapeLayer->setPath(path);
-
-  auto textLayer = TextLayer::Make();
-  textLayer->setText("style");
-  textLayer->setFont(Font(typeface, 100.f));
-  textLayer->setTextColor(Color::FromRGBA(192, 192, 192, 255));
-  textLayer->setLayerStyles({noise});
-  textLayer->setMatrix(Matrix::MakeTrans(50.f, 50.f));
-
-  shapeLayer->addChild(textLayer);
-  back->addChild(shapeLayer);
-  displayList->root()->addChild(back);
-
-  for (int i = 0; i < 8; i++) {
-    surface->getCanvas()->clear();
-    shapeLayer->setMatrix(Matrix::MakeTrans(static_cast<float>(i) * 20.f, 0.f));
-    displayList->render(surface.get());
-    EXPECT_TRUE(
-        Baseline::Compare(surface, "NoiseStyleTest/NoiseShapeText_frame" + std::to_string(i)));
-  }
-}
-
-static std::shared_ptr<Image> MakeNoiseTileImage(int tileSize) {
-  auto noiseShader = Shader::MakeFractalNoise(0.04f, 0.04f, 3, 42.0f);
-  if (noiseShader == nullptr) {
-    return nullptr;
-  }
-  PictureRecorder recorder = {};
-  auto recordCanvas = recorder.beginRecording();
-  Paint paint = {};
-  paint.setShader(noiseShader);
-  recordCanvas->drawRect(Rect::MakeWH(static_cast<float>(tileSize), static_cast<float>(tileSize)),
-                         paint);
-  auto picture = recorder.finishRecordingAsPicture();
-  if (picture == nullptr) {
-    return nullptr;
-  }
-  return Image::MakeFrom(std::move(picture), tileSize, tileSize);
-}
-
-TGFX_TEST(NoiseStyleTest, NoiseMovingRect) {
-  ContextScope scope;
-  auto context = scope.getContext();
-  ASSERT_TRUE(context != nullptr);
-  auto surface = Surface::Make(context, 1000, 1000);
-  ASSERT_TRUE(surface != nullptr);
-  auto displayList = std::make_unique<DisplayList>();
-
-  auto back = SolidLayer::Make();
-  back->setColor(Color::White());
-  back->setWidth(1000);
-  back->setHeight(1000);
-
-  auto parent = ShapeLayer::Make();
-  parent->setFillStyle(ShapeStyle::Make(Color::White()));
-  Path parentPath;
-  parentPath.addRect(Rect::MakeXYWH(50.f, 50.f, 900.f, 900.f));
-  parent->setPath(parentPath);
-
-  auto child = ShapeLayer::Make();
-  child->setFillStyle(ShapeStyle::Make(Color::White()));
-  auto noise = NoiseStyle::MakeMono(10.0f, 0.505f, Color::FromRGBA(0, 0, 0, 128), 42.0f);
-  child->setLayerStyles({noise});
-  parent->addChild(child);
-
-  back->addChild(parent);
-  displayList->root()->addChild(back);
-
-  for (int i = 0; i < 16; i++) {
-    surface->getCanvas()->clear();
-    Path childPath;
-    auto offsetX = -150.f + static_cast<float>(i) * 20.f;
-    childPath.addRect(Rect::MakeXYWH(offsetX, 50.f, 200.f, 200.f));
-    child->setPath(childPath);
-    displayList->render(surface.get());
-    EXPECT_TRUE(
-        Baseline::Compare(surface, "NoiseStyleTest/NoiseMovingRect_frame" + std::to_string(i)));
-  }
-}
-
-TGFX_TEST(NoiseStyleTest, NoiseMovingRectTiled) {
-  ContextScope scope;
-  auto context = scope.getContext();
-  ASSERT_TRUE(context != nullptr);
-  auto surface = Surface::Make(context, 1000, 1000);
-  ASSERT_TRUE(surface != nullptr);
-  auto displayList = std::make_unique<DisplayList>();
-  displayList->setRenderMode(RenderMode::Tiled);
-  displayList->setTileSize(64);
-
-  auto back = SolidLayer::Make();
-  back->setColor(Color::White());
-  back->setWidth(1000);
-  back->setHeight(1000);
-
-  auto parent = ShapeLayer::Make();
-  parent->setFillStyle(ShapeStyle::Make(Color::White()));
-  Path parentPath;
-  parentPath.addRect(Rect::MakeXYWH(50.f, 50.f, 900.f, 900.f));
-  parent->setPath(parentPath);
-
-  auto child = ShapeLayer::Make();
-  child->setFillStyle(ShapeStyle::Make(Color::White()));
-  auto noise = NoiseStyle::MakeMono(10.0f, 0.505f, Color::FromRGBA(0, 0, 0, 128), 42.0f);
-  child->setLayerStyles({noise});
-  parent->addChild(child);
-
-  back->addChild(parent);
-  displayList->root()->addChild(back);
-
-  for (int i = 0; i < 16; i++) {
-    surface->getCanvas()->clear();
-    Path childPath;
-    auto offsetX = -150.f + static_cast<float>(i) * 20.f;
-    childPath.addRect(Rect::MakeXYWH(offsetX, 50.f, 200.f, 200.f));
-    child->setPath(childPath);
-    displayList->render(surface.get());
-    EXPECT_TRUE(Baseline::Compare(surface,
-                                  "NoiseStyleTest/NoiseMovingRectTiled_frame" + std::to_string(i)));
-  }
-}
-
-TGFX_TEST(NoiseStyleTest, NoiseTileMode) {
-  ContextScope scope;
-  auto context = scope.getContext();
-  ASSERT_TRUE(context != nullptr);
-  auto surface = Surface::Make(context, 1000, 1000);
-  ASSERT_TRUE(surface != nullptr);
-  auto displayList = std::make_unique<DisplayList>();
-
-  auto back = SolidLayer::Make();
-  back->setColor(Color::White());
-  back->setWidth(1000);
-  back->setHeight(1000);
-  displayList->root()->addChild(back);
-
-  auto tileImage = MakeNoiseTileImage(100);
-  ASSERT_TRUE(tileImage != nullptr);
-
-  auto tileModes = {TileMode::Clamp, TileMode::Repeat, TileMode::Mirror, TileMode::Decal};
-  int col = 0;
-  for (auto mode : tileModes) {
-    auto pattern = ImagePattern::Make(tileImage, mode, mode);
-    pattern->setScaleMode(ScaleMode::None);
-    pattern->setMatrix(Matrix::MakeScale(0.5f));
-
-    auto imageShader = pattern->getShader();
-    auto shapeStyle = ShapeStyle::Make(imageShader);
-
-    auto shapeLayer = ShapeLayer::Make();
-    Path path = {};
-    path.addRect(Rect::MakeXYWH(0, 0, 200, 200));
-    shapeLayer->setPath(path);
-    shapeLayer->setFillStyle(shapeStyle);
-    shapeLayer->setMatrix(Matrix::MakeTrans(50.f + static_cast<float>(col) * 230.f, 50.f));
-
-    back->addChild(shapeLayer);
-    col++;
-  }
-
-  displayList->render(surface.get());
-  EXPECT_TRUE(Baseline::Compare(surface, "NoiseStyleTest/NoiseTileMode"));
-}
-
-TGFX_TEST(NoiseStyleTest, NoiseMovingRectWithDropShadow) {
-  ContextScope scope;
-  auto context = scope.getContext();
-  ASSERT_TRUE(context != nullptr);
-  auto surface = Surface::Make(context, 1000, 1000);
-  ASSERT_TRUE(surface != nullptr);
-  auto displayList = std::make_unique<DisplayList>();
-
-  auto back = SolidLayer::Make();
-  back->setColor(Color::White());
-  back->setWidth(1000);
-  back->setHeight(1000);
-
-  auto parent = ShapeLayer::Make();
-  parent->setFillStyle(ShapeStyle::Make(Color::White()));
-  Path parentPath;
-  parentPath.addRect(Rect::MakeXYWH(50.f, 50.f, 900.f, 900.f));
-  parent->setPath(parentPath);
-
-  auto child = ShapeLayer::Make();
-  child->setFillStyle(ShapeStyle::Make(Color::White()));
-  auto noise = NoiseStyle::MakeMono(10.0f, 0.505f, Color::FromRGBA(0, 0, 0, 128), 42.0f);
-  auto shadow = DropShadowStyle::Make(10.f, 10.f, 5.f, 5.f, Color::FromRGBA(0, 0, 0, 128), false);
-  child->setLayerStyles({noise, shadow});
-  parent->addChild(child);
-
-  back->addChild(parent);
-  displayList->root()->addChild(back);
-
-  for (int i = 0; i < 16; i++) {
-    surface->getCanvas()->clear();
-    Path childPath;
-    auto offsetX = -150.f + static_cast<float>(i) * 20.f;
-    childPath.addRect(Rect::MakeXYWH(offsetX, 50.f, 200.f, 200.f));
-    child->setPath(childPath);
-    displayList->render(surface.get());
-    EXPECT_TRUE(Baseline::Compare(
-        surface, "NoiseStyleTest/NoiseMovingRectWithDropShadow_frame" + std::to_string(i)));
-  }
-}
-
-}  // namespace tgfx
